+++ stereocam/capture_images.py
@@ -36,7 +36,7 @@ def capture_stereo(output_path="images", camera_number=None, width=4416, height=
         cam_available = detect_camera()
 
         # Selecting stereo camera
-        camera_number = detect_stereo(cam_available)
+        _, camera_number = detect_stereo(cam_available)
 
     # date
     ct = datetime.datetime.now()
@@ -149,6 +149,8 @@ def detect_stereo(cam_available, stereo_res=(1242, 4416)):
     # Using the last value of the cam_available list because it has a potential to be stereo
     stereo_index = cam_available.pop()
 
+    stereo_retval = True
+
     cap = cv2.VideoCapture(stereo_index)
 
     # Creating resolution
@@ -162,11 +164,12 @@ def detect_stereo(cam_available, stereo_res=(1242, 4416)):
         print('\033[92m' + 'Stereo possible')
     else:
         print('\033[93m' + 'May not be stereo. Please check manually and use correct index!')
+        stereo_retval = False
 
     cap.release()
     cv2.destroyAllWindows()
 
-    return stereo_index
+    return stereo_retval, stereo_index
 
 def initialise_camera(cap):
     """Initialises the camera for five iterations to avoid missing data.
@@ -188,7 +191,7 @@ def initialise_camera(cap):
             print('\033[91m' + "Camera failed to start!")
             break
 
-def record_stereo(output_path, camera_number, width, height):
+def record_stereo(output_path, camera_number, width, height, fps=10, save_images=False):
     """Records stereo videos.
     
     Parameters
@@ -219,7 +222,7 @@ def record_stereo(output_path, camera_number, width, height):
         cam_available = detect_camera()
 
         # Selecting stereo camera
-        camera_number = detect_stereo(cam_available)
+        _, camera_number = detect_stereo(cam_available)
 
     # date 
     ct = datetime.datetime.now()
@@ -245,9 +248,22 @@ def record_stereo(output_path, camera_number, width, height):
     initialise_camera(cap)
 
     # Create video writers for left and right cameras
-    left_writer = cv2.VideoWriter(os.path.join(path_left, 'left.avi'), cv2.VideoWriter_fourcc(*'XVID'), 30, (width // 2, height))
-    right_writer = cv2.VideoWriter(os.path.join(path_right, 'right.avi'), cv2.VideoWriter_fourcc(*'XVID'), 30, (width // 2, height))
+    left_writer = cv2.VideoWriter(os.path.join(path_left, 'left.avi'), cv2.VideoWriter_fourcc(*'XVID'), fps, (width // 2, height))
+    right_writer = cv2.VideoWriter(os.path.join(path_right, 'right.avi'), cv2.VideoWriter_fourcc(*'XVID'), fps, (width // 2, height))
 
+    
+    # Creates "images" directory inside the stereo left and right directory
+    if save_images:
+        left_images_path = os.path.join(path_left, "images")
+        right_images_path = os.path.join(path_right, "images")
+
+        if not os.path.exists(left_images_path):
+            os.makedirs(left_images_path)
+        if not os.path.exists(right_images_path):
+            os.makedirs(right_images_path)
+
+    # Counts the frame
+    frame_counter = 0
 
     while cap.isOpened():
         # Read frames from the stereo camera
@@ -261,6 +277,13 @@ def record_stereo(output_path, camera_number, width, height):
         left_frame = frame[:, :width // 2, :]
         right_frame = frame[:, width // 2:, :]
 
+        if save_images:
+            left_image_save_path = os.path.join(left_images_path, str(frame_counter) + ".png")
+            right_image_save_path = os.path.join(right_images_path, str(frame_counter) + ".png")
+
+            cv2.imwrite(left_image_save_path, left_frame)
+            cv2.imwrite(right_image_save_path, right_frame)
+
         # Write frames to the corresponding video writers
         left_writer.write(left_frame)
         right_writer.write(right_frame)
@@ -272,6 +295,8 @@ def record_stereo(output_path, camera_number, width, height):
         # Break the loop if 'q' is pressed
         if cv2.waitKey(1) & 0xFF == ord('q'):
             break
+        
+        frame_counter += 1
 
     # Release the video capture and video writers
     cap.release()
@@ -280,3 +305,19 @@ def record_stereo(output_path, camera_number, width, height):
 
     # Destroy all windows
     cv2.destroyAllWindows()
+
+def split_stereo(frame):
+    """Splits the stereo image into left and right frames.
+
+    Parameters
+    ----------
+    frame: numpy.ndarray
+        Stereo image frame.
+
+    """
+    height, width, _ = frame.shape 
+
+    left_frame = frame[:, :width // 2, :]
+    right_frame = frame[:, width // 2:, :]
+
+    return left_frame, right_frame