+++ CHANGELOG.rst
@@ -9,6 +9,19 @@ and this project adheres to `Semantic Versioning`_.
 .. _Keep a Changelog: https://keepachangelog.com/en/1.0.0/
 .. _Semantic Versioning: https://semver.org/spec/v2.0.0.html
 
+[Unreleased]
+------------
+
+Added
+^^^^^
+- Streamlit application added for SGBM tuning.
+- Weighted Least Square (WSL) filter added.
+
+Changed
+^^^^^^^
+- Detect stereo function returns stereo detection boolean flag along with camera number.
+
+
 [0.0.1] - 2025-01-13
 --------------------